Split Snapshot.swift by class
